Remove pepe
Even though the threshold is `0` on this one, I don't like it.

When this first came up a year ago, I thought "We can't throw out everything
terrible people use to terrible."

After the year we've had I no longer stand by the comment I made on #459.

I hope this doesn't spark debate.  In all honestly I don't even want to have
a long discussion about it.  I simply want to stand up and say I think there
is no value to this script, weighing that against the potential of how it may
make some feel or reflect on the spotbot community and our employer, it's time
to retire it.